vision/requirements.txt
ultralytics
fastapi
uvicorn[standard]
python-multipart
pydantic
opencv-python
python-dotenv